wtai/requirements.txt
anyio==4.8.0
certifi==2024.12.14
googletrans==4.0.2
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
protobuf==5.29.3
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
sniffio==1.3.1
typing_extensions==4.12.2
# Editable install with no version control (wtai==0.1.0)
-e /Users/rhaqim/Documents/Projects/worldtalker/wtai
